Fix for flask limiter 4.0 Fix reseting flask_limiter keys
